--- src/dcmReader/dcm_fixed_characteristic_map.py
@@ -1,10 +1,0 @@
-"""
-Definition of DCM fixed characteristic map
-"""
-from __future__ import annotations
-
-from dcmReader.dcm_characteristic_map import DcmCharacteristicMap
-
-
-class DcmFixedCharacteristicMap(DcmCharacteristicMap):
-    """Definition of a fixed characteristic map, derived from characteristic map"""
