--- src/dcmReader/dcm_group_characteristic_map.py
@@ -1,10 +1,0 @@
-"""
-Definition of DCM fixed characteristic map
-"""
-from __future__ import annotations
-
-from dcmReader.dcm_characteristic_map import DcmCharacteristicMap
-
-
-class DcmGroupCharacteristicMap(DcmCharacteristicMap):
-    """Definition of a group characteristic map, derived from characteristic map"""
